LORIS utilities package (PR 1) (#1374)
* decouple utils from lib

* loris utils package
