+++ pyproject.toml
@@ -27,6 +27,7 @@ dependencies = [
     "virtualenv",
     "loris-bids-reader @ {root:uri}/python/loris_bids_reader",
     "loris-eeg-chunker @ {root:uri}/python/loris_eeg_chunker",
+    "loris-utils       @ {root:uri}/python/loris_utils",
 ]
 
 [project.optional-dependencies]
@@ -72,7 +73,6 @@ include = [
     "python/lib/imaging_lib",
     "python/lib/import_bids_dataset",
     "python/lib/import_dicom_study",
-    "python/lib/util",
     "python/lib/bids.py",
     "python/lib/config.py",
     "python/lib/config_file.py",
@@ -83,6 +83,7 @@ include = [
     "python/scripts/import_dicom_study.py",
     "python/scripts/summarize_dicom_study.py",
     "python/loris_bids_reader",
+    "python/loris_utils",
 ]
 typeCheckingMode = "strict"
 reportMissingTypeStubs = "none"
