+++ python/lib/imaging_io.py
@@ -12,13 +12,13 @@ from lib.exitcode import COPY_FAILURE
 """Set of io functions."""
 
 
-@deprecated('Use `lib.logging` and `lib.util.fs` instead')
+@deprecated('Use `lib.logging` and `loris_utils.fs` instead')
 class ImagingIO:
     def __init__(self, log_obj, verbose):
         self.log_obj = log_obj
         self.verbose = verbose
 
-    @deprecated('Use `lib.util.fs.extract_archive` instead')
+    @deprecated('Use `loris_utils.fs.extract_archive` instead')
     def extract_archive(self, location, prefix, tmp_dir):
         """
         Extract Archive in the temporary directory
@@ -35,7 +35,7 @@ class ImagingIO:
         tar_file.close()
         return extract_location
 
-    @deprecated('Use `lib.util.fs.remove_directory` instead')
+    @deprecated('Use `shutil.rmtree` instead')
     def remove_dir(self, dir):
         """
         Removes a directory and its content
@@ -47,7 +47,7 @@ class ImagingIO:
             except PermissionError as err:
                 self.log_info(f"Could not delete {dir}. Error was: {err}", is_error=True, is_verbose=False)
 
-    @deprecated('Use `lib.util.fs.copy_file` instead')
+    @deprecated('Use `shutil.copytree` instead')
     def copy_file(self, old_file_path, new_file_path):
         """
         Move a file on the file system.
