+++ python/loris_utils/README.md
@@ -1,0 +1,3 @@
+# LORIS utilities
+
+This package contains a bunch of LORIS-agnostic utility functions and code to be used in LORIS code or related packages.
