+++ python/loris_utils/src/loris_utils/fs.py
@@ -1,18 +1,13 @@
 import os
 import re
-import shutil
 import tarfile
 import tempfile
 from collections.abc import Iterator
 from datetime import datetime
 from pathlib import Path
 
-import lib.exitcode
-from lib.env import Env
-from lib.logging import log_error_exit, log_verbose, log_warning
 
-
-def extract_archive(env: Env, tar_path: str, prefix: str, dir_path: str) -> str:
+def extract_archive(tar_path: str, prefix: str, dir_path: str) -> str:
     """
     Extract an archive in a new temporary directory inside the given directory and return
     the new directory location.
@@ -38,29 +33,6 @@ def iter_all_dir_files(dir_path: Path) -> Iterator[Path]:
             yield from iter_all_dir_files(item_path)
         elif item_path.is_file():
             yield item_path.relative_to(dir_path)
-
-
-def remove_directory(env: Env, path: str):
-    """
-    Delete a directory and its content.
-    """
-
-    if os.path.exists(path):
-        try:
-            shutil.rmtree(path)
-        except PermissionError as error:
-            log_warning(env, f"Could not delete {path}. Error was: {error}")
-
-
-def copy_file(env: Env, old_path: str, new_path: str):
-    """
-    Copy a file on the file system.
-    """
-
-    log_verbose(env, f"Moving {old_path} to {new_path}")
-    shutil.copytree(old_path, new_path, dirs_exist_ok=True)
-    if not os.path.exists(new_path):
-        log_error_exit(env, f"Could not copy {old_path} to {new_path}", lib.exitcode.COPY_FAILURE)
 
 
 def is_directory_empty(dir_path: str) -> bool:
