+++ python/loris_utils/src/loris_utils/iter.py
